move recipe format function to cake stuff
